First Winnable Version!
This is a huge milestone! It is now possible to actually play through the game
and eventually win, no matter how lacklustre the winning message. There is a
librarian in your room for debugging purposes.

Most of the change has been the introduction of the conversation system, which
uses a simple numerically based menu.

You now need to use "and" or "with" to combine items, but the items can have
spaces in the names. But not the words "and" or "with", obviously.